# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CollBar — Collective Bargaining Compensation Modeling API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /districts:
    get:
      operationId: listDistricts
      tags: [districts]
      summary: List all districts
      responses:
        "200":
          description: List of districts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/District"
    post:
      operationId: createDistrict
      tags: [districts]
      summary: Create a district
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDistrictRequest"
      responses:
        "201":
          description: Created district
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/District"
  /districts/{id}:
    get:
      operationId: getDistrict
      tags: [districts]
      summary: Get a district
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: District
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/District"
    put:
      operationId: updateDistrict
      tags: [districts]
      summary: Update a district
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDistrictRequest"
      responses:
        "200":
          description: Updated district
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/District"
  /bargaining-units:
    get:
      operationId: listBargainingUnits
      tags: [bargaining-units]
      summary: List bargaining units
      parameters:
        - name: districtId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of bargaining units
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BargainingUnit"
    post:
      operationId: createBargainingUnit
      tags: [bargaining-units]
      summary: Create a bargaining unit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBargainingUnitRequest"
      responses:
        "201":
          description: Created bargaining unit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BargainingUnit"
  /bargaining-units/{id}:
    get:
      operationId: getBargainingUnit
      tags: [bargaining-units]
      summary: Get a bargaining unit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Bargaining unit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BargainingUnit"
    put:
      operationId: updateBargainingUnit
      tags: [bargaining-units]
      summary: Update a bargaining unit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBargainingUnitRequest"
      responses:
        "200":
          description: Updated bargaining unit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BargainingUnit"
    delete:
      operationId: deleteBargainingUnit
      tags: [bargaining-units]
      summary: Delete a bargaining unit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /employees:
    get:
      operationId: listEmployees
      tags: [employees]
      summary: List employees
      parameters:
        - name: districtId
          in: query
          schema:
            type: string
        - name: bargainingUnitId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: contractYear
          in: query
          schema:
            type: number
        - name: page
          in: query
          schema:
            type: number
            default: 1
        - name: pageSize
          in: query
          schema:
            type: number
            default: 50
      responses:
        "200":
          description: Paginated list of employees
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeList"
    post:
      operationId: createEmployee
      tags: [employees]
      summary: Create an employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmployeeRequest"
      responses:
        "201":
          description: Created employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/retirement-eligible:
    get:
      operationId: listRetirementEligibleEmployees
      tags: [employees]
      summary: List retirement-eligible employees with incentive calculations
      parameters:
        - name: districtId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of retirement-eligible employees with calculations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RetirementEligibleEmployee"
  /employees/import:
    post:
      operationId: importEmployees
      tags: [employees]
      summary: Bulk import employees from CSV/Excel data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportEmployeesRequest"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
  /employees/{id}:
    get:
      operationId: getEmployee
      tags: [employees]
      summary: Get an employee with projections
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: scenarioId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Employee with projections
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeWithProjections"
    put:
      operationId: updateEmployee
      tags: [employees]
      summary: Update an employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEmployeeRequest"
      responses:
        "200":
          description: Updated employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    delete:
      operationId: deleteEmployee
      tags: [employees]
      summary: Delete an employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    District:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        state:
          type: string
        fiscalYearStart:
          type: string
        studentEnrollment:
          type: number
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - state
        - fiscalYearStart
        - createdAt
        - updatedAt
    CreateDistrictRequest:
      type: object
      properties:
        name:
          type: string
        state:
          type: string
        fiscalYearStart:
          type: string
        studentEnrollment:
          type: number
          nullable: true
      required:
        - name
    UpdateDistrictRequest:
      type: object
      properties:
        name:
          type: string
        state:
          type: string
        fiscalYearStart:
          type: string
        studentEnrollment:
          type: number
          nullable: true
    BargainingUnit:
      type: object
      properties:
        id:
          type: string
        districtId:
          type: string
        name:
          type: string
        code:
          type: string
        compensationType:
          type: string
          enum: [salary, hourly]
        retirementSystem:
          type: string
          enum: [TRS, IMRF, other]
        retirementEmployeeRate:
          type: string
        retirementEmployerRate:
          type: string
        retirementGrossUpRate:
          type: string
        ficaRate:
          type: string
        ficaExempt:
          type: boolean
        healthInsuranceSingleAnnual:
          type: string
        healthInsuranceFamilyAnnual:
          type: string
        dentalAnnual:
          type: string
        lifeInsuranceAnnual:
          type: string
        disabilityInsuranceAnnual:
          type: string
        hsaContributionSingle:
          type: string
        hsaContributionFamily:
          type: string
        workersCompRate:
          type: string
        contractStartDate:
          type: string
          nullable: true
        contractEndDate:
          type: string
          nullable: true
        contractYears:
          type: number
        displayOrder:
          type: number
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - districtId
        - name
        - code
        - compensationType
        - retirementSystem
        - contractYears
        - createdAt
        - updatedAt
    CreateBargainingUnitRequest:
      type: object
      properties:
        districtId:
          type: string
        name:
          type: string
        code:
          type: string
        compensationType:
          type: string
          enum: [salary, hourly]
        retirementSystem:
          type: string
          enum: [TRS, IMRF, other]
        retirementEmployeeRate:
          type: string
        retirementGrossUpRate:
          type: string
        ficaRate:
          type: string
        ficaExempt:
          type: boolean
        healthInsuranceSingleAnnual:
          type: string
        healthInsuranceFamilyAnnual:
          type: string
        hsaContributionSingle:
          type: string
        hsaContributionFamily:
          type: string
        workersCompRate:
          type: string
        contractStartDate:
          type: string
          nullable: true
        contractEndDate:
          type: string
          nullable: true
        contractYears:
          type: number
        displayOrder:
          type: number
      required:
        - districtId
        - name
        - code
        - compensationType
    UpdateBargainingUnitRequest:
      type: object
      properties:
        name:
          type: string
        retirementSystem:
          type: string
        retirementEmployeeRate:
          type: string
        retirementGrossUpRate:
          type: string
        ficaRate:
          type: string
        ficaExempt:
          type: boolean
        healthInsuranceSingleAnnual:
          type: string
        healthInsuranceFamilyAnnual:
          type: string
        hsaContributionSingle:
          type: string
        hsaContributionFamily:
          type: string
        workersCompRate:
          type: string
        contractStartDate:
          type: string
          nullable: true
        contractEndDate:
          type: string
          nullable: true
        contractYears:
          type: number
    Employee:
      type: object
      properties:
        id:
          type: string
        districtId:
          type: string
        bargainingUnitId:
          type: string
        employeeNumber:
          type: string
          nullable: true
        firstName:
          type: string
        lastName:
          type: string
        hireDate:
          type: string
          nullable: true
        birthDate:
          type: string
          nullable: true
        yearsInDistrict:
          type: number
        yearsTotalService:
          type: number
        compensationType:
          type: string
        currentLaneId:
          type: string
          nullable: true
        currentStep:
          type: number
          nullable: true
        currentHourlyCategoryId:
          type: string
          nullable: true
        currentHourlyRate:
          type: string
          nullable: true
        annualHours:
          type: string
          nullable: true
        currentAnnualSalary:
          type: string
        insuranceElection:
          type: string
        retirementEligible:
          type: boolean
        retirementPlan:
          type: string
        status:
          type: string
        contractYear:
          type: number
        bargainingUnitName:
          type: string
          nullable: true
        laneName:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - districtId
        - bargainingUnitId
        - firstName
        - lastName
        - currentAnnualSalary
        - insuranceElection
        - status
        - createdAt
        - updatedAt
    EmployeeList:
      type: object
      properties:
        employees:
          type: array
          items:
            $ref: "#/components/schemas/Employee"
        total:
          type: number
        page:
          type: number
        pageSize:
          type: number
      required:
        - employees
        - total
        - page
        - pageSize
    EmployeeWithProjections:
      allOf:
        - $ref: "#/components/schemas/Employee"
        - type: object
          properties:
            yearProjections:
              type: array
              items:
                $ref: "#/components/schemas/EmployeeYearRecord"
            retirementOptions:
              $ref: "#/components/schemas/RetirementOptions"
              nullable: true
    EmployeeYearRecord:
      type: object
      properties:
        contractYear:
          type: number
        yearLabel:
          type: string
        projectedStep:
          type: number
          nullable: true
        projectedLaneId:
          type: string
          nullable: true
        projectedLaneName:
          type: string
          nullable: true
        projectedHourlyRate:
          type: string
          nullable: true
        projectedBaseSalary:
          type: string
        retirementContribution:
          type: string
        ficaCost:
          type: string
        healthInsuranceCost:
          type: string
        otherBenefitsCost:
          type: string
        totalEmployerCost:
          type: string
        effectiveRate:
          type: string
          nullable: true
      required:
        - contractYear
        - projectedBaseSalary
        - retirementContribution
        - ficaCost
        - healthInsuranceCost
        - otherBenefitsCost
        - totalEmployerCost
    CreateEmployeeRequest:
      type: object
      properties:
        districtId:
          type: string
        bargainingUnitId:
          type: string
        employeeNumber:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        hireDate:
          type: string
        birthDate:
          type: string
        yearsInDistrict:
          type: number
        yearsTotalService:
          type: number
        compensationType:
          type: string
        currentLaneId:
          type: string
        currentStep:
          type: number
        currentHourlyCategoryId:
          type: string
        currentHourlyRate:
          type: string
        annualHours:
          type: string
        currentAnnualSalary:
          type: string
        insuranceElection:
          type: string
        retirementEligible:
          type: boolean
        retirementPlan:
          type: string
        status:
          type: string
        contractYear:
          type: number
      required:
        - districtId
        - bargainingUnitId
        - firstName
        - lastName
        - currentAnnualSalary
    UpdateEmployeeRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        yearsInDistrict:
          type: number
        yearsTotalService:
          type: number
        currentLaneId:
          type: string
        currentStep:
          type: number
        currentAnnualSalary:
          type: string
        currentHourlyRate:
          type: string
        annualHours:
          type: string
        insuranceElection:
          type: string
        retirementEligible:
          type: boolean
        retirementPlan:
          type: string
        retirementTargetYear:
          type: number
        status:
          type: string
        notes:
          type: string
    ImportEmployeesRequest:
      type: object
      properties:
        districtId:
          type: string
        bargainingUnitId:
          type: string
        contractYear:
          type: number
        incremental:
          type: boolean
        employees:
          type: array
          items:
            $ref: "#/components/schemas/CreateEmployeeRequest"
      required:
        - districtId
        - employees
    ImportResult:
      type: object
      properties:
        imported:
          type: number
        skipped:
          type: number
        errors:
          type: array
          items:
            type: object
            properties:
              row:
                type: number
              message:
                type: string
      required:
        - imported
        - skipped
        - errors
    RetirementEligibleEmployee:
      allOf:
        - $ref: "#/components/schemas/Employee"
        - type: object
          properties:
            retirementOptions:
              $ref: "#/components/schemas/RetirementOptions"
    RetirementOptions:
      type: object
      properties:
        option1:
          type: object
          nullable: true
        option2:
          type: object
          nullable: true
        option3:
          type: object
          nullable: true
...