# lib/api-zod/src/generated/types/employeeYearRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CollBar — Collective Bargaining Compensation Modeling API
 * OpenAPI spec version: 0.1.0
 */

export interface EmployeeYearRecord {
  contractYear: number;
  yearLabel?: string;
  projectedStep?: number | null;
  projectedLaneId?: string | null;
  projectedLaneName?: string | null;
  projectedHourlyRate?: string | null;
  projectedBaseSalary: string;
  retirementContribution: string;
  ficaCost: string;
  healthInsuranceCost: string;
  otherBenefitsCost: string;
  totalEmployerCost: string;
  effectiveRate?: string | null;
}